unify encryption
