+++ src-tauri/src/ethereum.rs
@@ -91,8 +91,6 @@ impl GethProcess {
 
         // Always kill any existing geth processes before starting
         // This ensures we don't have multiple instances running
-        println!("Cleaning up any existing geth processes...");
-
         // First try to stop via HTTP if it's running
         if self.is_running() {
             let _ = Command::new("curl")
@@ -130,7 +128,6 @@ impl GethProcess {
 
         // Final check - if still running, we have a problem
         if self.is_running() {
-            println!("Error: Could not stop existing geth process");
             // Try one more aggressive kill
             #[cfg(unix)]
             {
@@ -148,8 +145,6 @@ impl GethProcess {
                 );
             }
         }
-
-        println!("Geth cleanup complete, starting new instance...");
 
         // Use the GethDownloader to get the correct path
         let downloader = crate::geth_downloader::GethDownloader::new();
@@ -234,11 +229,8 @@ impl GethProcess {
 
         // Add miner address if provided
         if let Some(address) = miner_address {
-            println!("Setting miner.etherbase to: {}", address);
             // Set the etherbase (coinbase) for mining rewards
             cmd.arg("--miner.etherbase").arg(address);
-        } else {
-            println!("No miner address provided, starting without etherbase");
         }
 
         // Create log file for geth output
@@ -261,8 +253,6 @@ impl GethProcess {
     }
 
     pub fn stop(&mut self) -> Result<(), String> {
-        println!("Stopping geth process...");
-
         // First try to kill the tracked child process
         if let Some(mut child) = self.child.take() {
             // Try to kill the process
@@ -270,22 +260,17 @@ impl GethProcess {
                 Ok(_) => {
                     // Wait for the process to actually exit
                     let _ = child.wait();
-                    println!("Tracked geth process terminated successfully");
                 }
-                Err(e) => {
-                    println!("Failed to kill tracked geth process: {}", e);
+                Err(_) => {
+                    // Process was already dead or couldn't be killed
                 }
             }
-        } else {
-            println!("No tracked child process, will kill by pattern");
         }
 
         // Always kill any geth processes by name as a fallback
         // This handles orphaned processes
         #[cfg(unix)]
         {
-            println!("Killing any geth processes by pattern...");
-
             // Kill by process name
             let result = Command::new("pkill")
                 .arg("-9")
@@ -295,13 +280,11 @@ impl GethProcess {
 
             match result {
                 Ok(output) => {
-                    if output.status.success() {
-                        println!("Successfully killed geth processes");
-                    } else {
-                        println!("pkill returned non-zero (no processes found or error)");
-                    }
+                    // pkill completed
                 }
-                Err(e) => println!("Failed to run pkill: {}", e),
+                Err(e) => {
+                    // Failed to run pkill
+                },
             }
 
             // Also kill by port usage
