+++ src/lib/dht.ts
@@ -87,9 +87,6 @@ export class DhtService {
     // Use default bootstrap nodes if none provided
     if (bootstrapNodes.length === 0) {
       bootstrapNodes = DEFAULT_BOOTSTRAP_NODES;
-      console.log("Using default bootstrap nodes for network connectivity");
-    } else {
-      console.log(`Using ${bootstrapNodes.length} custom bootstrap nodes`);
     }
 
     try {
@@ -116,8 +113,6 @@ export class DhtService {
       const peerId = await invoke<string>("start_dht_node", payload);
       this.peerId = peerId;
       this.port = port;
-      console.log("DHT started with peer ID:", this.peerId);
-      console.log("Your multiaddr for others to connect:", this.getMultiaddr());
       return this.peerId;
     } catch (error) {
       console.error("Failed to start DHT:", error);
@@ -130,7 +125,6 @@ export class DhtService {
     try {
       await invoke("stop_dht_node");
       this.peerId = null;
-      console.log("DHT stopped");
     } catch (error) {
       console.error("Failed to stop DHT:", error);
       throw error;
@@ -149,7 +143,6 @@ export class DhtService {
         fileSize: metadata.fileSize,
         mimeType: metadata.mimeType,
       });
-      console.log("Published file metadata:", metadata.fileHash);
     } catch (error) {
       console.error("Failed to publish file:", error);
       throw error;
@@ -163,7 +156,6 @@ export class DhtService {
 
     try {
       await invoke("search_file_metadata", { fileHash, timeoutMs: 0 });
-      console.log("Searching for file:", fileHash);
     } catch (error) {
       console.error("Failed to search file:", error);
       throw error;
@@ -182,7 +174,6 @@ export class DhtService {
 
     try {
       await invoke("connect_to_peer", { peerAddress });
-      console.log("Connecting to peer:", peerAddress);
     } catch (error) {
       console.error("Failed to connect to peer:", error);
       throw error;
