+++ src/pages/Mining.svelte
@@ -254,9 +254,6 @@
 
     await checkGethStatus()
     await updateNetworkStats()
-    try {
-      // seenHashes initialization removed - no longer used
-    } catch{} 
     // If mining is already active from before, restore session and update stats
     if ($miningState.isMining) {
       // Restore session start time if it exists
@@ -326,13 +323,7 @@
           if (hashRateFromLogs > 0) {
             // Use actual hash rate from logs
             $miningState.hashRate = formatHashRate(hashRateFromLogs)
-            
-            console.log('From mining stats - blocks:', blocksFound);
-            console.log('Current blocksFound before:', $miningState.blocksFound);
-            
             $miningState.blocksFound = blocksFound;
-            
-            console.log('blocksFound after:', $miningState.blocksFound);
           
             
           } else if ($miningState.activeThreads > 0) {
@@ -430,13 +421,8 @@
       
       if (results[4] !== undefined) {
         const blocksMined = results[4] as number;
-        
-        console.log('Backend returned blocks:', blocksMined);
-        console.log('Current blocksFound before update:', $miningState.blocksFound);
-        
+
         $miningState.blocksFound = blocksMined;
-        
-        console.log('blocksFound after update:', $miningState.blocksFound);
       }
     }
   } catch (e) {
@@ -1752,21 +1738,4 @@
         </div>
       </div>
     </div>
-  {/if} 
-<div class="balance-info" style="background: #f5f5f5; padding: 15px; margin: 10px 0; border-radius: 8px;">
-  <h3>Account Information</h3>
-  <div style="display: flex; gap: 30px;">
-    <div>
-      <strong>Current Balance:</strong> {($miningState.totalRewards || 0).toFixed(2)} Chiral
-    </div>
-    <div>
-      <strong>Blocks Found:</strong> {$miningState.blocksFound} blocks
-    </div>
-    <div>
-      <strong>Mining Status:</strong> {$miningState.isMining ? 'Active' : 'Stopped'}
-    </div>
-    <div>
-      <strong>Account:</strong> {$etcAccount ? `${$etcAccount.address.slice(0, 10)}...${$etcAccount.address.slice(-8)}` : 'No Account'}
-    </div>
-  </div>
-</div>
+  {/if}
