+++ src/pages/Network.svelte
@@ -248,12 +248,11 @@
       try {
         backendPeerId = await invoke<string | null>('get_dht_peer_id')
       } catch (error) {
-        console.log('Failed to check backend DHT status:', error)
+        // Failed to check backend DHT status
       }
       
       if (backendPeerId) {
         // DHT is already running in backend, sync the frontend state immediately
-        console.log('DHT already running in backend with peer ID:', backendPeerId)
         dhtPeerId = backendPeerId
         dhtService.setPeerId(backendPeerId) // Update frontend service state
         dhtEvents = [...dhtEvents, `✓ DHT already running with peer ID: ${backendPeerId.slice(0, 16)}...`]
@@ -330,7 +329,6 @@
       // Try to connect to bootstrap nodes
       let connectionSuccessful = false
       if (DEFAULT_BOOTSTRAP_NODES.length > 0) {
-        console.log('Attempting to connect to bootstrap nodes:', DEFAULT_BOOTSTRAP_NODES)
         dhtEvents = [...dhtEvents, `[Attempt ${connectionAttempts}] Connecting to ${DEFAULT_BOOTSTRAP_NODES.length} bootstrap node(s)...`]
         
         // Add another small delay to show the connection attempt
@@ -339,7 +337,6 @@
         try {
           // Try connecting to the first available bootstrap node
           await dhtService.connectPeer(DEFAULT_BOOTSTRAP_NODES[0])
-          console.log('Connection initiated to bootstrap nodes')
           connectionSuccessful = true
           dhtEvents = [...dhtEvents, `✓ Connection initiated to bootstrap nodes (waiting for handshake...)`]
           
@@ -574,11 +571,9 @@
         isInitiator: true,
         onMessage: (data) => {
           showToast('Received from peer: ' + data, 'info');
-          console.log('Received from peer:', data);
         },
         onConnectionStateChange: (state) => {
           showToast('WebRTC state: ' + state, 'info');
-          console.log('WebRTC connection state:', state);
           
           if (state === 'connected') {
             showToast('Successfully connected to peer!', 'success');
@@ -719,7 +714,6 @@
 
   async function startGethNode() {
     if (!isTauri) {
-      console.log('Cannot start Chiral Node in web mode - desktop app required')
       return
     }
     
@@ -745,7 +739,6 @@
 
   async function stopGethNode() {
     if (!isTauri) {
-      console.log('Cannot stop Chiral Node in web mode - desktop app required')
       return
     }
     
@@ -804,11 +797,9 @@
         signalingConnected = true;
         signaling.peers.subscribe(peers => {
           discoveredPeers = peers;
-          console.log('Updated discovered peers:', peers);
         });
-        console.log('Signaling service initialized successfully');
       } catch (error) {
-        console.log('Signaling service not available (DHT not running) - this is normal');
+        // Signaling service not available (DHT not running) - this is normal
         signalingConnected = false;
       }
