+++ src/pages/Upload.svelte
@@ -182,7 +182,7 @@
                 try {
                   existingVersions = await invoke('get_file_versions_by_name', { fileName: file.name }) as any[];
                 } catch (versionError) {
-                  console.log('No existing versions found for', file.name);
+                  // No existing versions found
                 }
 
                 // Use fileService.uploadFile method for File objects with versioning
@@ -200,7 +200,7 @@
                   const updatedVersions = await invoke('get_file_versions_by_name', { fileName: file.name }) as any[];
                   versionInfo = updatedVersions.find(v => v.file_hash === hash);
                 } catch (versionError) {
-                  console.log('Could not get version info for uploaded file');
+                  // Could not get version info for uploaded file
                 }
 
                 const isNewVersion = existingVersions.length > 0;
@@ -240,7 +240,6 @@
                     createdAt: Date.now(),
                     isEncrypted: false
                   });
-                  console.log('Dropped file published to DHT:', hash);
                 } catch (publishError) {
                   console.warn('Failed to publish dropped file to DHT:', publishError);
                 }
@@ -316,7 +315,6 @@
         // Stop publishing file to DHT network
         try {
           await invoke('stop_publishing_file', { fileHash });
-          console.log('File unpublished from DHT:', fileHash);
         } catch (unpublishError) {
           console.warn('Failed to unpublish file from DHT:', unpublishError);
         }
@@ -343,7 +341,7 @@
         try {
           existingVersions = await invoke('get_file_versions_by_name', { fileName }) as any[];
         } catch (versionError) {
-          console.log('No existing versions found for', fileName);
+          // No existing versions found
         }
 
         // Use versioned upload
@@ -394,7 +392,6 @@
         // Publish file metadata to DHT network for discovery
         try {
           await dhtService.publishFile(metadata);
-          console.log('File published to DHT:', metadata.file_hash);
         } catch (publishError) {
           console.warn('Failed to publish file to DHT:', publishError);
           // Don't show error to user as upload succeeded, just DHT publishing failed
